requirements.txt
Flask==3.0.0
flask-cors==4.0.0
rembg==2.0.50
Pillow==10.1.0
numpy==1.24.3
onnxruntime==1.16.3
gunicorn==21.2.0